Change slideshow wait position
Using this with a rather slow wifi, and big pictures, I often got black screens for most of the time, even with 15 seconds slideshow. Thus I am suggesting to move the waiting loop (while) for the slideshow from the end of the for loop to the beginning so that Kodi gets the most time to load the image.
Those changes were tested and seemed to work fine. I don't think there are any drawback even for fast slideshows.